Improve address handling abstraction (#896)
* Add distributed.comm.inproc draft

* Replace close_request hack with a custom peekable queue

* Add deserialization tests

* Add multi-thread test for inproc queues

* Add inproc tests to test_core

* Refactor address manipulation functions in a separate d.comm.addressing module

* Disable zmq test when zmq isn't explicitly enabled

* Fix loads() bug with hand-created Serialized object

* Move deserialization helper into distributed.protocol

* Remove deserialize argument from Comm.read()

* Get rid of some port numbers

* Better abstract address handling

* Add a __hash__ to Serialize

* Remove duplicated code due to merge gotcha

* Improve comms documentation

* Appease flake8

* Fix typos @mrocklin

+++ distributed/bokeh/application.py
@@ -31,8 +31,12 @@ logging_level = logging_names[logging_level.upper()]
 
 
 class BokehWebInterface(object):
-    def __init__(self, host='127.0.0.1', http_port=9786, tcp_port=8786,
-                 bokeh_port=8787, bokeh_whitelist=[], log_level=logging_level,
+
+    process = None
+
+    def __init__(self, host='127.0.0.1', http_port=9786, bokeh_port=8787,
+                 scheduler_address='tcp://127.0.0.1:8786',
+                 bokeh_whitelist=[], log_level=logging_level,
                  show=False, prefix=None, use_xheaders=False, quiet=True):
         self.port = bokeh_port
         ip = socket.gethostbyname(host)
@@ -73,7 +77,7 @@ class BokehWebInterface(object):
 
         bokeh_options = {'host': host,
                          'http-port': http_port,
-                         'tcp-port': tcp_port,
+                         'scheduler-address': scheduler_address,
                          'bokeh-port': bokeh_port}
 
         args.extend(['--args'] + list(map(str, concat(bokeh_options.items()))))
@@ -94,7 +98,7 @@ class BokehWebInterface(object):
                          % (ip, bokeh_port))
 
     def close(self, join=True, timeout=None):
-        if self.process.poll() is None:
+        if self.process is not None and self.process.poll() is None:
             self.process.terminate()
 
     def __del__(self):
